Split trigonometric/hyperbolic routines from math.cc
