+++ chainerx_cc/chainerx/cuda/cuda_device/hyperbolic.cu
@@ -1,0 +1,34 @@
+#include "chainerx/cuda/cuda_device.h"
+
+#include <cmath>
+#include <cstdint>
+
+#include <cuda_runtime.h>
+
+#include "chainerx/array.h"
+#include "chainerx/cuda/cuda_runtime.h"
+#include "chainerx/cuda/cuda_set_device_scope.h"
+#include "chainerx/cuda/elementwise.cuh"
+#include "chainerx/cuda/kernel_regist.h"
+#include "chainerx/cuda/numeric.cuh"
+#include "chainerx/device.h"
+#include "chainerx/dtype.h"
+#include "chainerx/kernels/hyperbolic.h"
+#include "chainerx/numeric.h"
+#include "chainerx/scalar.h"
+
+namespace chainerx {
+namespace cuda {
+namespace {
+
+CHAINERX_CUDA_REGISTER_ELTWISE_FLOAT_UNARY_KERNEL(SinhKernel, { out = cuda::Sinh(x); });
+
+CHAINERX_CUDA_REGISTER_ELTWISE_FLOAT_UNARY_KERNEL(CoshKernel, { out = cuda::Cosh(x); });
+
+CHAINERX_CUDA_REGISTER_ELTWISE_FLOAT_UNARY_KERNEL(ArcsinhKernel, { out = cuda::Arcsinh(x); });
+
+CHAINERX_CUDA_REGISTER_ELTWISE_FLOAT_UNARY_KERNEL(ArccoshKernel, { out = cuda::Arccosh(x); });
+
+}  // namespace
+}  // namespace cuda
+}  // namespace chainerx
