+++ chainerx_cc/chainerx/kernels/CMakeLists.txt
@@ -2,6 +2,7 @@ install(FILES
     binary.h
     connection.h
     creation.h
+    hyperbolic.h
     indexing.h
     linalg.h
     logic.h
@@ -10,5 +11,6 @@ install(FILES
     normalization.h
     pooling.h
     sorting.h
+    trigonometric.h
     DESTINATION include/chainerx/kernels
     )
