+++ chainerx_cc/chainerx/kernels/math.h
@@ -128,86 +128,9 @@ public:
     virtual void Call(const Array& x, const Array& out) = 0;
 };
 
-class SinKernel : public Kernel {
-public:
-    static const char* name() { return "Sin"; }
-
-    virtual void Call(const Array& x, const Array& out) = 0;
-};
-
-class CosKernel : public Kernel {
-public:
-    static const char* name() { return "Cos"; }
-
-    virtual void Call(const Array& x, const Array& out) = 0;
-};
-
-class TanKernel : public Kernel {
-public:
-    static const char* name() { return "Tan"; }
-
-    virtual void Call(const Array& x, const Array& out) = 0;
-};
-
-class SinhKernel : public Kernel {
-public:
-    static const char* name() { return "Sinh"; }
-
-    virtual void Call(const Array& x, const Array& out) = 0;
-};
-
-class CoshKernel : public Kernel {
-public:
-    static const char* name() { return "Cosh"; }
-
-    virtual void Call(const Array& x, const Array& out) = 0;
-};
-
 class TanhKernel : public Kernel {
 public:
     static const char* name() { return "Tanh"; }
-
-    virtual void Call(const Array& x, const Array& out) = 0;
-};
-
-class ArcsinKernel : public Kernel {
-public:
-    static const char* name() { return "Arcsin"; }
-
-    virtual void Call(const Array& x, const Array& out) = 0;
-};
-
-class ArccosKernel : public Kernel {
-public:
-    static const char* name() { return "Arccos"; }
-
-    virtual void Call(const Array& x, const Array& out) = 0;
-};
-
-class ArctanKernel : public Kernel {
-public:
-    static const char* name() { return "Arctan"; }
-
-    virtual void Call(const Array& x, const Array& out) = 0;
-};
-
-class Arctan2Kernel : public Kernel {
-public:
-    static const char* name() { return "Arctan2"; }
-
-    virtual void Call(const Array& x1, const Array& x2, const Array& out) = 0;
-};
-
-class ArcsinhKernel : public Kernel {
-public:
-    static const char* name() { return "Archsinh"; }
-
-    virtual void Call(const Array& x, const Array& out) = 0;
-};
-
-class ArccoshKernel : public Kernel {
-public:
-    static const char* name() { return "Arccosh"; }
 
     virtual void Call(const Array& x, const Array& out) = 0;
 };
