+++ chainerx_cc/chainerx/native/CMakeLists.txt
@@ -23,6 +23,7 @@ add_library(chainerx_native STATIC
     native_device/dot.cc
     native_device/exp_log.cc
     native_device/fill.cc
+    native_device/hyperbolic.cc
     native_device/indexing.cc
     native_device/memory.cc
     native_device/misc.cc
