+++ chainerx_cc/chainerx/routines/hyperbolic.h
@@ -1,0 +1,16 @@
+#pragma once
+
+#include "chainerx/array.h"
+#include "chainerx/scalar.h"
+
+namespace chainerx {
+
+Array Sinh(const Array& x);
+
+Array Cosh(const Array& x);
+
+Array Arcsinh(const Array& x);
+
+Array Arccosh(const Array& x);
+
+}  // namespace chainerx
